requirements.txt
absl-py==2.0.0
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
attrs==23.1.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
distro==1.8.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.45.1
fsspec==2023.10.0
gast==0.5.4
ginza==5.1.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.25.2
idna==3.5
importlib-metadata==6.8.0
ja-ginza==5.1.3
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
keras==2.15.0
kiwisolver==1.4.5
langcodes==3.3.0
libclang==16.0.6
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.10.8
ml-dtypes==0.2.0
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
openai==1.3.5
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
pathy==0.10.3
Pillow==10.1.0
plac==1.4.1
preshed==3.0.9
protobuf==3.20.3
psutil==5.9.6
py-cpuinfo==9.0.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rpds-py==0.13.1
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
streamlit==1.28.2
SudachiDict-core==20230927
SudachiPy==0.6.7
sympy==1.12
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-hub==0.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thinc==8.1.12
thop==0.1.1.post2209072238
toml==0.10.2
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ultralytics==8.0.217
urllib3==2.1.0
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
weasel==0.3.4
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
